refactor: Product jdbc -> jpa 로 변경
