--- src/main/java/kitchenpos/menu/domain/MenuProductRepository.java
@@ -1,10 +1,0 @@
-package kitchenpos.menu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
-
-    List<MenuProduct> findAllByMenu(Long id);
-}
