+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,7 +1,8 @@
 package kitchenpos.application;
 
-import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -20,7 +21,7 @@ import static org.mockito.BDDMockito.given;
 @ExtendWith(MockitoExtension.class)
 public class ProductServiceTest {
     @Mock
-    private ProductDao productDao;
+    private ProductRepository productRepository;
 
     @InjectMocks
     private ProductService productService;
@@ -37,7 +38,7 @@ public class ProductServiceTest {
     @DisplayName("상품을 이름, 가격으로 등록")
     @Test
     void create() {
-        given(productDao.save(강정치킨)).willReturn(강정치킨);
+        given(productRepository.save(강정치킨)).willReturn(강정치킨);
 
         Product createProduct = productService.create(강정치킨);
 
@@ -47,7 +48,7 @@ public class ProductServiceTest {
     @DisplayName("상품 목록 조회")
     @Test
     void list() {
-        given(productDao.findAll()).willReturn(Arrays.asList(강정치킨));
+        given(productRepository.findAll()).willReturn(Arrays.asList(강정치킨));
 
         List<Product> products = productService.list();
